+++ Rdmp.Core/Providers/ExamplePluginCohortCompilerUI.cs
@@ -4,47 +4,54 @@
 // RDMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
+using MapsDirectlyToDatabaseTable;
 using Rdmp.Core;
 using Rdmp.Core.CohortCreation.Execution;
 using Rdmp.Core.CommandExecution;
 using Rdmp.Core.Curation.Data.Aggregation;
 
-namespace Rdmp.UI.PluginChildProvision
+namespace Rdmp.Core.Providers
 {
     class ExamplePluginCohortCompilerUI : PluginUserInterface
     {
         private ExamplePluginCohortCompiler compiler;
 
-        public ExamplePluginCohortCompilerUI(IBasicActivateItems activator) :base(activator)
+        public ExamplePluginCohortCompilerUI(IBasicActivateItems activator) : base(activator)
         {
             compiler = new ExamplePluginCohortCompiler();
         }
 
-        public override bool CustomActivate(AggregateConfiguration ac)
+        public override bool CustomActivate(IMapsDirectlyToDatabaseTable o)
         {
+            // we only care about responding to opening AggregateConfiguration objects
+            if (o is not AggregateConfiguration ac)
+            {
+                return false;
+            }
+
             // If the aggregate being edited is one of the ones that our API handles
-            if(compiler.ShouldRun(ac))
+            if (compiler.ShouldRun(ac))
             {
                 int number;
 
                 // Look at the Description property for a number (your API could use a complex XML or YAML syntax if you want)
                 // You could also store some info at the Catalogue level for reuse in other Aggregates
 
-                if(!int.TryParse(ac.Description, out number))
+                if (!int.TryParse(ac.Description, out number))
                 {
                     number = 5;
                 }
 
                 // Launch a UI that prompts a new value to be entered
-                if(BasicActivator.TypeText("Generate random CHIs","Number of Chis:", 100, number.ToString(),out string result,false))
+                if (BasicActivator.TypeText("Generate random CHIs", "Number of Chis:", 100, number.ToString(), out string result, false))
                 {
-                    if(int.TryParse(result, out int newCount))
+                    if (int.TryParse(result, out int newCount))
                     {
                         ac.Description = newCount.ToString();
                         ac.SaveToDatabase();
                     }
                 }
-                
+
                 // we handled this, don't launch the default user interface
                 return true;
             }